slack-sync/requirements.txt
slack-sdk>=3.23.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
flask>=3.0.0
functions-framework>=3.5.0
